[Minor] Types for Alert component
